Added spring security & jwt

+++ model/src/main/java/com/epam/esm/dao/TagDao.java
@@ -3,7 +3,6 @@ package com.epam.esm.dao;
 import com.epam.esm.entity.QueryParameters;
 import com.epam.esm.entity.Tag;
 
-import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
@@ -11,15 +10,15 @@ public interface TagDao {
 
     List<Tag> findByParameters(QueryParameters parameters);
 
-    Optional<Tag> findById(BigInteger id);
+    Optional<Tag> findById(Long id);
 
-    BigInteger deleteById(BigInteger id);
+    Long deleteById(Long id);
 
     Tag save(Tag tag);
 
     Optional<Tag> findByName(String name);
 
-    Tag findMostUsedTag(BigInteger idUser);
+    Tag findMostUsedTag(Long idUser);
 
     long count();
 }
