+++ model/src/main/java/com/epam/esm/dao/impl/GiftCertificateTagDaoImpl.java
@@ -9,7 +9,6 @@ import javax.persistence.EntityManagerFactory;
 import javax.persistence.EntityNotFoundException;
 import javax.persistence.PersistenceContext;
 import javax.persistence.PersistenceUnit;
-import java.math.BigInteger;
 import java.util.Optional;
 
 @Repository
@@ -26,7 +25,7 @@ public class GiftCertificateTagDaoImpl implements GiftCertificateTagDao {
     }
 
     @Override
-    public Optional<GiftCertificateTag> findById(BigInteger id) {
+    public Optional<GiftCertificateTag> findById(Long id) {
         GiftCertificateTag foundGiftCertificateTag = entityManager.find(GiftCertificateTag.class, id);
         return Optional.ofNullable(foundGiftCertificateTag);
     }
@@ -42,7 +41,7 @@ public class GiftCertificateTagDaoImpl implements GiftCertificateTagDao {
     }
 
     @Override
-    public BigInteger deleteById(BigInteger id) {
+    public Long deleteById(Long id) {
         try {
             GiftCertificateTag giftCertificateTag = entityManager.find(GiftCertificateTag.class, id);
             if (giftCertificateTag != null) {
