+++ model/src/main/java/com/epam/esm/dao/impl/TagDaoImpl.java
@@ -23,7 +23,6 @@ import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Predicate;
 import javax.persistence.criteria.Root;
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -84,13 +83,13 @@ public class TagDaoImpl implements TagDao {
     }
 
     @Override
-    public Optional<Tag> findById(BigInteger id) {
+    public Optional<Tag> findById(Long id) {
         Tag foundTag = entityManager.find(Tag.class, id);
         return Optional.ofNullable(foundTag);
     }
 
     @Override
-    public BigInteger deleteById(BigInteger id) {
+    public Long deleteById(Long id) {
         Tag tag = entityManager.find(Tag.class, id);
         if (tag != null) {
             if (tag.getGiftCertificateTags() != null) {
@@ -128,7 +127,7 @@ public class TagDaoImpl implements TagDao {
     }
 
     @Override
-    public Tag findMostUsedTag(BigInteger idUser) {
+    public Tag findMostUsedTag(Long idUser) {
         try {
             return (Tag) entityManager.createNativeQuery(QUERY_MOST_USED_TAG, Tag.class)
                     .setParameter(ID_USER_PARAMETER, idUser)
