+++ model/src/main/java/com/epam/esm/entity/GiftCertificate.java
@@ -18,7 +18,6 @@ import javax.persistence.ManyToMany;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Iterator;
@@ -33,7 +32,7 @@ public class GiftCertificate {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private BigInteger id;
+    private Long id;
     @Column(name = "name")
     private String name;
     @Column(name = "description")
@@ -65,7 +64,7 @@ public class GiftCertificate {
     )
     private List<Order> orders;
 
-    public GiftCertificate(BigInteger id, String name, String description, BigDecimal price, LocalDateTime createDate, LocalDateTime lastUpdateDate, int duration, List<Tag> tags) {
+    public GiftCertificate(Long id, String name, String description, BigDecimal price, LocalDateTime createDate, LocalDateTime lastUpdateDate, int duration, List<Tag> tags) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -87,7 +86,7 @@ public class GiftCertificate {
         this.duration = duration;
     }
 
-    public GiftCertificate(BigInteger id, String name, String description, LocalDateTime lastUpdateDate) {
+    public GiftCertificate(Long id, String name, String description, LocalDateTime lastUpdateDate) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -97,11 +96,11 @@ public class GiftCertificate {
     public GiftCertificate() {
     }
 
-    public BigInteger getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(BigInteger id) {
+    public void setId(Long id) {
         this.id = id;
     }
