+++ model/src/main/java/com/epam/esm/entity/audit/OrderAudit.java
@@ -7,7 +7,6 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Table;
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.Objects;
 
@@ -16,7 +15,7 @@ import java.util.Objects;
 public class OrderAudit extends AbstractEntityAudit {
 
     @Column(name = "id_row")
-    private BigInteger idRow;
+    private Long idRow;
     @Column(name = "cost")
     private BigDecimal cost;
     @Column(name = "purchase_date")
@@ -34,11 +33,11 @@ public class OrderAudit extends AbstractEntityAudit {
         this.purchaseDate = order.getPurchaseDate();
     }
 
-    public BigInteger getIdRow() {
+    public Long getIdRow() {
         return idRow;
     }
 
-    public void setIdRow(BigInteger id) {
+    public void setIdRow(Long id) {
         this.idRow = id;
     }
