+++ service/src/main/java/com/epam/esm/dto/TagDto.java
@@ -1,12 +1,11 @@
 package com.epam.esm.dto;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.util.Objects;
 
 public class TagDto implements Serializable {
 
-    private BigInteger id;
+    private Long id;
     private String name;
 
     public TagDto() {
@@ -16,16 +15,16 @@ public class TagDto implements Serializable {
         this.name = name;
     }
 
-    public TagDto(BigInteger id, String name) {
+    public TagDto(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public BigInteger getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(BigInteger id) {
+    public void setId(Long id) {
         this.id = id;
     }
